Working connection to postgres via JPA via Hibernate via H2

+++ src/main/java/pl/edu/agh/softwarestudio/angel/location/LocationRepo.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LocationRepo extends JpaRepository<Location, Long> {
+public interface LocationRepo extends JpaRepository<Location, Integer> {
 }
